% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_noise}
\alias{add_noise}
\title{Add noise to boundary conditions}
\usage{
add_noise(bc)
}
\arguments{
\item{bc}{vector or matrix; meteorological boundary conditions}
}
\value{
vector or matrix; meteorological boundary conditions with noise
}
\description{
Adds random noise to data.
}
