% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_dens}
\alias{calc_dens}
\title{Calculate water density from temperature}
\usage{
calc_dens(wtemp)
}
\arguments{
\item{wtemp}{vector or matrix; Water temperatures}
}
\value{
vector or matrix; Water densities in kg/m3
}
\description{
Calculate water density from water temperature using the formula from (Millero & Poisson, 1981).
}
