% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{configure_from_ler}
\alias{configure_from_ler}
\title{Configure model with LakeEnsemblR data}
\usage{
configure_from_ler(config_file = "LakeEnsemblR.yaml", folder = ".")
}
\arguments{
\item{bc}{meteorological boundary conditions: day, shortwave radiation, air temperature, dew
point temperature, wind speed and wind shear stress}

\item{params}{configuration parameters}

\item{ini}{vector of the initial water temperatures of the epilimnion and hypolimnion}

\item{times}{vector of time information}
}
\value{
matrix of simulated water temperatures in the pilimnion and hypolimnion
}
\description{
Use existing configuration data from LakeEnsemblR to prepare configuration files for the model
}
