% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{EnKF}
\alias{EnKF}
\title{wrapper for running EnKF}
\usage{
EnKF(
  n_en = 100,
  start,
  stop,
  time_step = "days",
  obs_file = "A_EcoForecast/Data/lake_c_data.rds",
  driver_file = "A_EcoForecast/Data/lake_c_data.rds",
  n_states_est = 1,
  n_params_est = 1,
  n_params_obs = 0,
  decay_init = 0.005,
  obs_cv = 0.1,
  param_cv = 0.25,
  driver_cv = c(0.2, 0.2, 0.2),
  init_cond_cv = 0.1
)
}
\arguments{
\item{n_en}{number of model ensembles}

\item{start}{start date of model run}

\item{stop}{date of model run}

\item{time_step}{model time step, defaults to days}

\item{obs_file}{observation file}

\item{driver_file}{driver data file}

\item{n_states_est}{number of states we're estimating}

\item{n_params_est}{number of parameters we're estimating}

\item{n_params_obs}{number of parameters for which we have observations}

\item{decay_init}{initial decay rate of DOC}

\item{obs_cv}{coefficient of variation of observations}

\item{param_cv}{coefficient of variation of parameters}

\item{driver_cv}{coefficient of variation of driver data for DOC Load, Discharge out, and Lake volume, respectively}

\item{init_cond_cv}{initial condition CV (what we're )}
}
\description{
wrapper for running EnKF
}
