% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_drivers}
\alias{get_drivers}
\title{matrix for holding driver data}
\usage{
get_drivers(
  drivers_df,
  model_dates,
  n_drivers,
  driver_colnames,
  driver_cv,
  n_step,
  n_en
)
}
\arguments{
\item{drivers_df}{dataframe which holds all the driver data}

\item{model_dates}{dates for model run}

\item{n_drivers}{number of model drivers}

\item{driver_colnames}{column names of the drivers in the driver dataframe}

\item{driver_cv}{coefficient of variation for each driver data}

\item{n_step}{number of model timesteps}

\item{n_en}{number of ensembles}
}
\description{
matrix for holding driver data
}
