% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_model_dates}
\alias{get_model_dates}
\title{retreive the model time steps based on start and stop dates and time step}
\usage{
get_model_dates(model_start, model_stop, time_step = "days")
}
\arguments{
\item{model_start}{model start date in date class}

\item{model_stop}{model stop date in date class}

\item{time_step}{model time step, defaults to daily timestep}
}
\description{
retreive the model time steps based on start and stop dates and time step
}
