% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_obs_error_matrix}
\alias{get_obs_error_matrix}
\title{observation error matrix, should be a square matrix where
  col & row = the number of states and params for which you have observations}
\usage{
get_obs_error_matrix(n_states, n_params_obs, n_step, state_sd, param_sd)
}
\arguments{
\item{n_states}{number of states we're updating in data assimilation routine}

\item{n_step}{number of model timesteps}

\item{state_sd}{vector of state observation standard deviation; assuming sd is constant through time}

\item{param_sd}{vector of parmaeter observation standard deviation; assuming sd is constant through time}

\item{n_param_obs}{number of parameters for which we have observations}
}
\description{
observation error matrix, should be a square matrix where
  col & row = the number of states and params for which you have observations
}
