% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_obs_matrix}
\alias{get_obs_matrix}
\title{turn observation dataframe into matrix}
\usage{
get_obs_matrix(obs_df, model_dates, n_step, n_states)
}
\arguments{
\item{obs_df}{observation data frame}

\item{model_dates}{dates over which you're modeling}

\item{n_step}{number of model time steps}

\item{n_states}{number of states we're updating in data assimilation routine}
}
\description{
turn observation dataframe into matrix
}
