% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_Y_vector}
\alias{get_Y_vector}
\title{vector for holding states and parameters for updating}
\usage{
get_Y_vector(n_states, n_params_est, n_step, n_en)
}
\arguments{
\item{n_states}{number of states we're updating in data assimilation routine}

\item{n_params_est}{number of parameters we're calibrating}

\item{n_step}{number of model timesteps}

\item{n_en}{number of ensembles}
}
\description{
vector for holding states and parameters for updating
}
