% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{initialize_Y}
\alias{initialize_Y}
\title{initialize Y vector with draws from distribution of obs}
\usage{
initialize_Y(
  Y,
  obs,
  init_params,
  n_states_est,
  n_params_est,
  n_params_obs,
  n_step,
  n_en,
  state_sd,
  param_sd
)
}
\arguments{
\item{Y}{Y vector}

\item{obs}{observation matrix}
}
\description{
initialize Y vector with draws from distribution of obs
}
