% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{kalman_filtering}
\alias{kalman_filtering}
\title{Smoothing of time series}
\usage{
kalman_filtering(time, series)
}
\arguments{
\item{bc}{vector or matrix; meteorological boundary conditions}
}
\value{
vector or matrix; meteorological boundary conditions with noise
}
\description{
Filters the signal using a Kalman filter
}
