% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run_model}
\alias{run_model}
\title{Simple two-layer water temperature model}
\usage{
run_model(bc, params, ini, times, ice = FALSE)
}
\arguments{
\item{bc}{meteorological boundary conditions: day, shortwave radiation, air temperature, dew
point temperature, wind speed and wind shear stress}

\item{params}{configuration parameters}

\item{ini}{vector of the initial water temperatures of the epilimnion and hypolimnion}

\item{times}{vector of time information}

\item{ice}{boolean, if TRUE the model will approximate ice on/off set, otherwise no ice}
}
\value{
matrix of simulated water temperatures in the pilimnion and hypolimnion
}
\description{
Runs a simple two-layer water temperature model in dependence of meteorological drivers and
lake setup/configuration. The model automatically calculates summer stratification onset and
offset
}
