% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run_npz_model}
\alias{run_npz_model}
\title{Simple two-layer water temperature, oxygen, phytoplankton, zooplankton and nutrient model}
\usage{
run_npz_model(bc, params, ini, times, ice = FALSE, new.diagn = TRUE)
}
\arguments{
\item{bc}{meteorological boundary conditions: day, shortwave radiation, air temperature, dew
point temperature, wind speed and wind shear stress}

\item{params}{configuration parameters}

\item{ini}{vector of the initial water temperatures of the epilimnion and hypolimnion}

\item{times}{vector of time information}

\item{ice}{boolean, if TRUE the model will approximate ice on/off set, otherwise no ice}
}
\value{
matrix of simulated water temperature and oxygen mass in the epilimnion and hypolimnion
}
\description{
Runs a simple two-layer water NPZ model in dependence of meteorological 
drivers and lake setup/configuration. The model automatically calculates summer stratification 
onset and offset
}
