% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{TwoLayer}
\alias{TwoLayer}
\title{Extract time and space information}
\usage{
TwoLayer(t, y, parms)
}
\arguments{
\item{t}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{y}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{parns}{matrix; Water temperatures (rows correspond to time, cols to depth)}
}
\value{
list of datetimes and depths
}
\description{
Extracts time (from date column) and space (aka depth) information
}
