% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{configure_from_ler}
\alias{configure_from_ler}
\title{Configure model with LakeEnsemblR data}
\usage{
configure_from_ler(config_file = "LakeEnsemblR.yaml", folder = ".")
}
\arguments{
\item{config_file}{string of your LER configuration file}

\item{folder}{folder path to the configuration file}
}
\value{
vector of parameters to run thermod
}
\description{
Use existing configuration data from LakeEnsemblR to prepare configuration files for the model
}
