% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run_model}
\alias{run_model}
\title{Extract time and space information}
\usage{
run_model(modelfunc = "TwoLayer", bc, params, ini, times)
}
\arguments{
\item{t}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{y}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{parns}{matrix; Water temperatures (rows correspond to time, cols to depth)}
}
\value{
list of datetimes and depths
}
\description{
Extracts time (from date column) and space (aka depth) information
}
